Send back error if user signs up with existing username
